Phase 3: Update the caller path and truncation note in get_address_info
